change handel fake data to create multi case in test case
